+++ MDATest/ICLTest/TestNoteRemoteness.cs
@@ -1,7 +1,4 @@
 ﻿using System;
-using System.Collections.Generic;
-using System.Linq;
-using System.Text;
 using Microsoft.VisualStudio.TestTools.UnitTesting;
 using MDA.OIP.BorodaDivider;
 using MDA.OIP.ScoreModel;
@@ -18,11 +15,11 @@ namespace MDATest.ICLTest
             Fmotiv fmotiv1 = new Fmotiv(0, "ПМТ");
             Fmotiv fmotiv2 = new Fmotiv(1, "ПМТ");
 
-            fmotiv1.NoteList.Add(new Note(new Pitch(3, 'E', 0), new Duration(1, 4, false, 512), false, Tie.None));
-            fmotiv1.NoteList.Add(new Note(new Pitch(3, 'A', 0), new Duration(1, 4, false, 512), false, Tie.None));
+            fmotiv1.NoteList.Add(new Note(new Pitch(3, 'E', 0), new Duration(1, 4, false, 512), false, (int)Tie.None));
+            fmotiv1.NoteList.Add(new Note(new Pitch(3, 'A', 0), new Duration(1, 4, false, 512), false, (int)Tie.None));
 
-            fmotiv2.NoteList.Add(new Note(new Pitch(3, 'E', 0), new Duration(1, 4, false, 512), false, Tie.None));
-            fmotiv2.NoteList.Add(new Note(new Pitch(3, 'A', 0), new Duration(1, 4, false, 512), false, Tie.None));
+            fmotiv2.NoteList.Add(new Note(new Pitch(3, 'E', 0), new Duration(1, 4, false, 512), false, (int)Tie.None));
+            fmotiv2.NoteList.Add(new Note(new Pitch(3, 'A', 0), new Duration(1, 4, false, 512), false, (int)Tie.None));
 
             // записываем ф-мотивы в цепь ф-мотивов, которая будет сравниваться с получившейся
             FmotivChain fmchain1 = new FmotivChain();
@@ -37,12 +34,12 @@ namespace MDATest.ICLTest
             Fmotiv fmotiv1 = new Fmotiv(0, "ПМТ");
             Fmotiv fmotiv2 = new Fmotiv(1, "ПМТ");
 
-            fmotiv1.NoteList.Add(new Note(new Pitch(3, 'E', 0), new Duration(1, 4, false, 512), false, Tie.None));
-            fmotiv1.NoteList.Add(new Note(new Pitch(3, 'A', 0), new Duration(1, 4, false, 512), false, Tie.None));
+            fmotiv1.NoteList.Add(new Note(new Pitch(3, 'E', 0), new Duration(1, 4, false, 512), false, (int)Tie.None));
+            fmotiv1.NoteList.Add(new Note(new Pitch(3, 'A', 0), new Duration(1, 4, false, 512), false, (int)Tie.None));
 
-            fmotiv2.NoteList.Add(new Note(new Pitch(3, 'E', 0), new Duration(1, 4, false, 512), false, Tie.None));
-            fmotiv2.NoteList.Add(new Note(new Pitch(3, 'A', 0), new Duration(1, 4, false, 512), false, Tie.None));
-            fmotiv2.NoteList.Add(new Note(null, new Duration(1, 4, false, 512), false, Tie.None));
+            fmotiv2.NoteList.Add(new Note(new Pitch(3, 'E', 0), new Duration(1, 4, false, 512), false, (int)Tie.None));
+            fmotiv2.NoteList.Add(new Note(new Pitch(3, 'A', 0), new Duration(1, 4, false, 512), false, (int)Tie.None));
+            fmotiv2.NoteList.Add(new Note(null, new Duration(1, 4, false, 512), false, (int)Tie.None));
 
             // записываем ф-мотивы в цепь ф-мотивов, которая будет сравниваться с получившейся
             FmotivChain fmchain1 = new FmotivChain();
@@ -57,12 +54,12 @@ namespace MDATest.ICLTest
             Fmotiv fmotiv1 = new Fmotiv(0, "ПМТ");
             Fmotiv fmotiv2 = new Fmotiv(1, "ПМТ");
 
-            fmotiv1.NoteList.Add(new Note(new Pitch(3, 'E', 0), new Duration(1, 4, false, 512), false, Tie.None));
-            fmotiv1.NoteList.Add(new Note(new Pitch(3, 'A', 0), new Duration(1, 2, false, 512), false, Tie.None));
+            fmotiv1.NoteList.Add(new Note(new Pitch(3, 'E', 0), new Duration(1, 4, false, 512), false, (int)Tie.None));
+            fmotiv1.NoteList.Add(new Note(new Pitch(3, 'A', 0), new Duration(1, 2, false, 512), false, (int)Tie.None));
 
-            fmotiv2.NoteList.Add(new Note(new Pitch(3, 'E', 0), new Duration(1, 4, false, 512), false, Tie.None));
-            fmotiv2.NoteList.Add(new Note(new Pitch(3, 'A', 0), new Duration(1, 4, false, 512), false, Tie.Start));
-            fmotiv2.NoteList.Add(new Note(new Pitch(3, 'A', 0), new Duration(1, 4, false, 512), false, Tie.Stop));
+            fmotiv2.NoteList.Add(new Note(new Pitch(3, 'E', 0), new Duration(1, 4, false, 512), false, (int)Tie.None));
+            fmotiv2.NoteList.Add(new Note(new Pitch(3, 'A', 0), new Duration(1, 4, false, 512), false, (int)Tie.Start));
+            fmotiv2.NoteList.Add(new Note(new Pitch(3, 'A', 0), new Duration(1, 4, false, 512), false, (int)Tie.Stop));
             
             // записываем ф-мотивы в цепь ф-мотивов, которая будет сравниваться с получившейся
             FmotivChain fmchain1 = new FmotivChain();
@@ -78,11 +75,11 @@ namespace MDATest.ICLTest
             Fmotiv fmotiv1 = new Fmotiv(0, "ПМТ");
             Fmotiv fmotiv2 = new Fmotiv(1, "ПМТ");
 
-            fmotiv1.NoteList.Add(new Note(new Pitch(2, 'E', 0), new Duration(1, 4, false, 512), false, Tie.None));
-            fmotiv1.NoteList.Add(new Note(new Pitch(3, 'A', 0), new Duration(1, 2, false, 512), false, Tie.None));
+            fmotiv1.NoteList.Add(new Note(new Pitch(2, 'E', 0), new Duration(1, 4, false, 512), false, (int)Tie.None));
+            fmotiv1.NoteList.Add(new Note(new Pitch(3, 'A', 0), new Duration(1, 2, false, 512), false, (int)Tie.None));
 
-            fmotiv1.NoteList.Add(new Note(new Pitch(3, 'E', 0), new Duration(1, 4, false, 512), false, Tie.None));
-            fmotiv1.NoteList.Add(new Note(new Pitch(4, 'A', 0), new Duration(1, 2, false, 512), false, Tie.None));
+            fmotiv1.NoteList.Add(new Note(new Pitch(3, 'E', 0), new Duration(1, 4, false, 512), false, (int)Tie.None));
+            fmotiv1.NoteList.Add(new Note(new Pitch(4, 'A', 0), new Duration(1, 2, false, 512), false, (int)Tie.None));
 
             // записываем ф-мотивы в цепь ф-мотивов, которая будет сравниваться с получившейся
             FmotivChain fmchain1 = new FmotivChain();
